Add bull and replace queue
