+++ package.json
@@ -57,7 +57,9 @@
     "@feathersjs/socketio": "3.2.2",
     "async": "^3.2.0",
     "axios": "^0.21.1",
+    "bee-queue": "^1.3.1",
     "bignumber.js": "^8.1.1",
+    "bull": "^3.22.0",
     "cli-progress": "3.8.2",
     "colors": "1.4.0",
     "compression": "^1.7.0",
